requirements.txt
azure-core==1.35.0
azure-data-tables==12.4.0
azure-functions==1.18.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.3.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
schedule==1.2.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
